réalisation de l'update de la parti commentaire

+++ resources/views/articles/commentaire.blade.php
@@ -6,30 +6,147 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>CRUD IN LARAVEL 10</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-  </head>
+<style>
+    .navbar-nav{
+    width: 100%;
+}
+
+@media(min-width:568px){
+    .end{
+        margin-left: auto;
+    }
+}
+
+@media(max-width:768px){
+    #post{
+        width: 100%;
+    }
+}
+#clicked{
+    padding-top: 1px;
+    padding-bottom: 1px;
+    text-align: center;
+    width: 100%;
+    background-color: #ecb21f;
+    border-color: #a88734 #9c7e31 #846a29;
+    color: black;
+    border-width: 1px;
+    border-style: solid;
+    border-radius: 13px; 
+}
+
+#profile{
+    background-color: unset;
+    
+} 
+
+#post{
+    margin: 10px;
+    padding: 6px;
+    padding-top: 2px;
+    padding-bottom: 2px;
+    text-align: center;
+    background-color: #ecb21f;
+    border-color: #a88734 #9c7e31 #846a29;
+    color: black;
+    border-width: 1px;
+    border-style: solid;
+    border-radius: 13px;
+    width: 50%;
+}
+
+body{
+    background-color: white;
+}
+
+#nav-items li a,#profile{
+    text-decoration: none;
+    color: rgb(224, 219, 219);
+    background-color: black;
+}
+
+
+.comments{
+    margin-top: 5%;
+    margin-left: 20px;
+}
+
+.darker{
+    border: 1px solid #ecb21f;
+    background-color: black;
+    float: right;
+    border-radius: 5px;
+    padding-left: 40px;
+    padding-right: 30px;
+    padding-top: 10px;
+}
+
+.comment{
+    border: 1px solid black;
+    background-color: #3093BA
+    float: left;
+    border-radius: 5px;
+    padding-left: 40px;
+    padding-right: 30px;
+    padding-top: 10px;
+    
+}
+.comment h4,.comment span,.darker h4,.darker span{
+    display: inline;
+}
+
+.comment p,.comment span,.darker p,.darker span{
+    color: rgb(184, 183, 183);
+}
+
+h1,h4{
+    color: white;
+    font-weight: bold;
+}
+label{
+    color: rgb(212, 208, 208);
+}
+
+#align-form{
+    margin-top: 20px;
+}
+.form-group p a{
+    color: white;
+}
+
+#checkbx{
+    background-color: black;
+}
+
+#darker img{
+    margin-right: 15px;
+    position: static;
+}
+
+.form-group input,.form-group textarea{
+    background-color: white;
+    border: 1px solid rgba(16, 46, 46, 1);
+    border-radius: 12px;
+}
+
+form{
+    border: 1px solid rgba(16, 46, 46, 1);
+    background-color: #3093BA;
+    border-radius: 5px;
+    padding: 20px;
+ }</style>  
+</head>
   <body>
 
   <nav class="navbar navbar-expand-sm navbar-dark">
-    <img src="https://i.imgur.com/CFpa3nK.jpg" width="20" height="20" class="d-inline-block align-top rounded-circle" alt=""> 
-    <a class="navbar-brand ml-2" href="#" data-abc="true">Rib Simpson</a>
-    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarColor02" aria-controls="navbarColor02" aria-expanded="false" aria-label="Toggle navigation"> 
-        <span class="navbar-toggler-icon"></span> 
-    </button>
-    <div class="end">
-        <div class="collapse navbar-collapse" id="navbarColor02">
-            <ul class="navbar-nav">
-                <li class="nav-item"> <a class="nav-link" href="#" data-abc="true">Work</a> </li>
-                <li class="nav-item"> <a class="nav-link" href="#" data-abc="true">Capabilities</a> </li>
-                <li class="nav-item "> <a class="nav-link" href="#" data-abc="true">Articles</a> </li>
-                <li class="nav-item active"> <a class="nav-link mt-2" href="#" data-abc="true" id="clicked">Contact<span class="sr-only">(current)</span></a> </li>
-            </ul>        
-        </div>
-    </div>    
+
 </nav>
 <!-- Main Body -->
 <section>
    
     <div class="container">
+        <a href="/" class="btn btn-info">Retour</a>
+
         <div class="row">
             <div class="col-sm-5 col-md-6 col-12 pb-4">
                 <h1>Commentaires</h1>
@@ -40,9 +157,10 @@
                     <span>- 20 October, 2018</span>
                     <br>
                     <p>{{$commentaire->contenu}}</p>
+                    <a href="/update-commentaire/{{$commentaire->id}}" class="btn btn-info">Modifier</a>
+                <a onclick="return confirm('Confirmer la suppression')" href="/delete-commentaire/{{$commentaire->id}}" class="btn btn-danger">Supprimer</a>
                 </div>
-                <a href="/update-commentaire/{{$commentaire->id}}" class="btn btn-info">Modifier</a>
-                <a href="/delete-commentaire/{{$commentaire->id}}" class="btn btn-danger">Supprimer</a>
+                
 
 
                 @endforeach
@@ -66,9 +184,9 @@
                       @endforeach
                     </ul>
                     <div class="form-group">
-                        <h4>Leave a comment</h4>
+                        <h4>Lessez nous un message !</h4>
                         <label for="message">Message</label>
-                        <textarea name="contenu" id=""msg cols="30" rows="5" class="form-control" style="background-color: black;" name="contenu"></textarea>
+                        <textarea name="contenu" id=""msg cols="30" rows="5" class="form-control" style="background-color: white;" name="contenu"></textarea>
                     </div>
                     <div class="form-group">
                         <label for="name">Nom</label>
@@ -76,12 +194,9 @@
                     </div>
                     <div class="form-group">
                         <label for="email">Date</label>
-                        <span>- {{ $commentaire->created_at->format('d M, Y') }}</span>
                         <input type="date" name="date_heure_creation" id="email" class="form-control">
                     </div>
                     <div class="form-group">
-                        <label for="email">Article_id</label>
-                        {{-- <input type="number" name="article_id" id="email" class="form-control">  --}}
                         <input type="text" name="article_id" style="display: none;"  value="{{$article->id}}">
 
                     </div>
